manager: Improve documentation
"Singleton" has a lot of bad reputation. Let's just avoid it.
Strictly speaking, this isn't a singleton class. A user can
always do g_object_new() to create as many instances and it's
not enforced to have one and only one instance.

Design wise, it might be not good to create an object accessible
to every other class. Well, this is a really small application.
May be we can consider to change this later.

The current approach makes it hard to have an owner for the object.
That means an implicit destruction happens on program exit, which
is bad. May be should design this to make #GnApplication the owner,
Still a singleton instance passing the object to others as needed.